Clean up setup.rb. Revive Request Fields
According to lsoren, setup.rb in templates/rest/fulldoc/html was
basically a copy-paste job from the default template in 0.6.8.  This is
an attempt at paring down the redundancies that may or may not have been
causing version incompatitibilities.

It also turns out that 'Request Fields' make sense when the only
acceptable is an xml or json string.